Refactor JAX build wheel rule and add wheel_library targets.
This change is a part of the initiative to test the JAX wheels in the presubmit properly.

The current setup is designed for postsubmit only, it consists of running two commands for producing the wheels (`bazel build` and `bazel run`), then launching docker, installing the wheels in venv, and then running bazel tests with `bazel test --//jax:build_jaxlib=false`.

The new JAX wheel build rule produces the wheel in the Build phase using `bazel build` command only. That means that the JAX wheel targets can be added as dependencies in other targets in Build phase.

This is a pre-requisite for running `bazel test --//jax:build_jaxlib=false` using one step only, without the need to build the wheels separately.

The list of the changes:
1) JAX wheel build rule verifies that `--@local_config_cuda//cuda:include_cuda_libs=false` during the wheel build. There is a way to pass the restriction by providing `--@local_config_cuda//cuda:override_include_cuda_libs=true`.

2) The wheel in the output of the build rule always has version `0.0.0`. This is a dummy version that allows installing the wheel via `pip install` or hermetic Python. The version value in the wheel name doesn't affect the actual wheel content.

3) To add the real version in the wheel names, the following commands should be executed:

   ```
   bazel run <flags> -- //jaxlib/tools:rename_jaxlib_wheel

   bazel run <flags> -- //jaxlib/tools:rename_jax_cuda_plugin_wheel

   bazel run <flags> -- //jaxlib/tools:rename_jax_cuda_pjrt_wheel
   ```

   By default the renamed wheels will be put in `dist` folder inside the workspace.
   To override the path, one should provide the flag `--output_path` in the command line script arguments.

PiperOrigin-RevId: 691109911